:recycle:  refactor: bulk insert 위치 및 코드 수정 [#34]
