+++ waggle-api/src/main/java/com/trip/api/chat/repository/ChatRoomMemberRepository.java
@@ -2,6 +2,8 @@ package com.trip.api.chat.repository;
 
 import com.trip.api.chat.entity.ChatRoomMember;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
+@Repository
 public interface ChatRoomMemberRepository extends JpaRepository<ChatRoomMember, Long> {
 }
